+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Each section shall contain a list of action items of the following format: `<bri
 
 ## Unreleased
 
+- Batch file (`.ci/asciidoc-converter/build_standard_debug.bat`) to aid troubleshooting document processing.
+
 ### Added
 
 ### Changes